Fixes #6287: Remove V1 API entirely.
This commit removes the existence of the V1 API by moving some controllers
to pure V2 API controllers and removing other, unused API controllers
and their corresponding tests.
