--- app/controllers/katello/api/v1/filters_controller.rb
@@ -1,23 +1,0 @@
-#
-# Katello Organization actions
-# Copyright 2014 Red Hat, Inc.
-#
-# This software is licensed to you under the GNU General Public License,
-# version 2 (GPLv2). There is NO WARRANTY for this software, express or
-# implied, including the implied warranties of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
-# along with this software; if not, see
-# http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
-#
-# Red Hat trademarks are not licensed under GPLv2. No permission is
-# granted to use or replicate Red Hat trademarks that are incorporated
-# in this software or its documentation.
-#
-
-module Katello
-class Api::V1::FiltersController < Api::V1::ApiController
-
-  # TODO: remove me
-
-end
-end
