--- app/controllers/katello/api/v1/uebercerts_controller.rb
@@ -1,31 +1,0 @@
-#
-# Copyright 2014 Red Hat, Inc.
-#
-# This software is licensed to you under the GNU General Public
-# License as published by the Free Software Foundation; either version
-# 2 of the License (GPLv2) or (at your option) any later version.
-# There is NO WARRANTY for this software, express or implied,
-# including the implied warranties of MERCHANTABILITY,
-# NON-INFRINGEMENT, or FITNESS FOR A PARTICULAR PURPOSE. You should
-# have received a copy of GPLv2 along with this software; if not, see
-# http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
-
-module Katello
-class Api::V1::UebercertsController < Api::V1::ApiController
-  before_filter :find_organization, :only => [:show]
-
-  before_filter :authorize
-
-  def rules
-    read_test = lambda { @organization.readable? }
-    { :show => read_test }
-  end
-
-  api :GET, "/organizations/:organization_id/uebercert", N_("Show an ueber certificate for an organization")
-  param :regenerate, :bool, :desc => N_("When set to 'True' certificate will be re-issued")
-  def show
-    @organization.generate_debug_cert if (params[:regenerate] || '').downcase == 'true'
-    respond :resource => @organization.debug_cert
-  end
-end
-end
