+++ app/controllers/katello/api/v2/uebercerts_controller.rb
@@ -11,13 +11,20 @@
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
 
 module Katello
-class Api::V2::UebercertsController < Api::V1::UebercertsController
+class Api::V2::UebercertsController < Api::V2::ApiController
 
-  include Api::V2::Rendering
+  before_filter :find_organization, :only => [:show]
 
   resource_description do
     api_version 'v2'
     api_base_url "#{Katello.config.url_prefix}/api"
+  end
+
+  api :GET, "/organizations/:organization_id/uebercert", N_("Show an ueber certificate for an organization")
+  param :regenerate, :bool, :desc => N_("When set to 'True' certificate will be re-issued")
+  def show
+    @organization.generate_debug_cert if (params[:regenerate] || '').downcase == 'true'
+    respond :resource => @organization.debug_cert
   end
 
 end
