--- app/lib/katello/api/version1.rb
@@ -1,9 +1,0 @@
-module Katello
-module Api
-  module Version1
-    def api_version
-      '1'
-    end
-  end
-end
-end
