+++ lib/katello/engine.rb
@@ -124,7 +124,6 @@ module Katello
       # similar strucuture and if the Foreman files are loaded first,
       # autoloading doesn't work.
       require_dependency "#{Katello::Engine.root}/app/controllers/katello/api/api_controller"
-      require_dependency "#{Katello::Engine.root}/app/controllers/katello/api/v1/api_controller"
       require_dependency "#{Katello::Engine.root}/app/controllers/katello/api/v2/api_controller"
       ::PuppetClassImporter.send :include, Katello::Services::PuppetClassImporterExtensions
     end
